cleanup transaction code
