+++ testing/ef-tests/src/lib.rs
@@ -36,11 +36,11 @@ use zeth_primitives::{
     block::Header,
     ethers::from_ethers_h160,
     keccak::keccak,
-    signature::TxSignature,
     transactions::{
         ethereum::{
             EthereumTxEssence, TransactionKind, TxEssenceEip1559, TxEssenceEip2930, TxEssenceLegacy,
         },
+        signature::TxSignature,
         EthereumTransaction,
     },
     trie::{self, MptNode, MptNodeData, StateAccount},
@@ -51,7 +51,6 @@ use zeth_primitives::{
 use crate::ethers::{get_state_update_proofs, TestProvider};
 
 pub mod ethers;
-
 pub mod ethtests;
 
 pub mod guests {
